+++ packages/app/src/DBSearchPage.tsx
@@ -956,6 +956,7 @@ function DBSearchPage() {
           displayedColumns,
           toggleColumn,
           generateSearchUrl,
+          dbSqlRowTableConfig,
         }}
       >
         {searchedSource && (