+++ packages/app/src/components/PageHeader.tsx
@@ -1,0 +1,5 @@
+import styles from './PageHeader.module.scss';
+
+export const PageHeader = ({ children }: { children: React.ReactNode }) => {
+  return <div className={styles.header}> {children}</div>;
+};