+++ packages/common-utils/src/types.ts
@@ -396,7 +396,15 @@ export enum SourceKind {
   Log = 'log',
   Trace = 'trace',
   Session = 'session',
+  Metric = 'metric',
 }
+
+export enum MetricsDataType {
+  Gauge = 'gauge',
+  Histogram = 'histogram',
+  Sum = 'sum',
+}
+
 export const SourceSchema = z.object({
   from: z.object({
     databaseName: z.string(),
@@ -440,6 +448,15 @@ export const SourceSchema = z.object({
   statusCodeExpression: z.string().optional(),
   statusMessageExpression: z.string().optional(),
   logSourceId: z.string().optional(),
+
+  // OTEL Metrics
+  metricTables: z
+    .object({
+      [MetricsDataType.Gauge]: z.string(),
+      [MetricsDataType.Histogram]: z.string(),
+      [MetricsDataType.Sum]: z.string(),
+    })
+    .optional(),
 });
 
 export type TSource = z.infer<typeof SourceSchema>;