+++ packages/app/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @hyperdx/app
 
+## 2.0.0-beta.10
+
+### Patch Changes
+
+- 459267a: feat: introduce session table model form
+
+## 2.0.0-beta.0
+
+### Major Changes
+
+- bumps to v2 beta
+
+### Patch Changes
+
+- b3f3151: Allow to create Slack Webhooks from Team Settings page
+- b690db8: Introduce event panel overview tab
+
 ## 1.9.0
 
 ### Minor Changes